added StateChangeable contract to InputWidget
