+++ android/android-declarative/src/main/java/br/com/zup/beagle/widget/interfaces/StateChangeable.kt
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package br.com.zup.beagle.interfaces
+package br.com.zup.beagle.widget.interfaces
 
-import br.com.zup.beagle.state.Observable
+import br.com.zup.beagle.widget.state.Observable
 
-interface Observer<T> {
-    fun update(o: Observable<T>, arg: T)
+interface StateChangeable {
+    fun getState(): Observable<WidgetState>
 }
+
+data class WidgetState(val value: Any)
