+++ android/sample/src/main/java/br/com/zup/beagle/sample/widgets/SampleTextField.kt
@@ -18,22 +18,15 @@ package br.com.zup.beagle.sample.widgets
 
 import android.content.Context
 import android.widget.EditText
-import br.com.zup.beagle.annotation.RegisterWidget
-import br.com.zup.beagle.interfaces.StateChangeable
-import br.com.zup.beagle.interfaces.WidgetState
-import br.com.zup.beagle.state.Observable
-import br.com.zup.beagle.widget.form.InputWidget
 import androidx.core.widget.doOnTextChanged
+import br.com.zup.beagle.annotation.RegisterWidget
+import br.com.zup.beagle.widget.form.InputWidget
 
 @RegisterWidget
-class SampleTextField(val placeholder: String) : InputWidget(), StateChangeable {
+class SampleTextField(private val placeholder: String) : InputWidget() {
 
     @Transient
     private lateinit var textFieldView: EditText
-    @Transient
-    private val stateObservable = Observable<WidgetState>()
-
-    override fun getState(): Observable<WidgetState> = stateObservable
 
     override fun getValue(): Any = textFieldView.text.toString()
 
@@ -46,7 +39,7 @@ class SampleTextField(val placeholder: String) : InputWidget(), StateChangeable 
         textFieldView.hint = placeholder
         textFieldView.isSingleLine = true
         this.doOnTextChanged { text, _, _, _ ->
-            stateObservable.notifyObservers(WidgetState(text.toString()))
+            notifyObservers(text.toString())
         }
     }
 }
