+++ android/sample/src/main/java/br/com/zup/beagle/sample/widgets/TextField.kt
@@ -22,10 +22,7 @@ import android.text.InputType
 import android.widget.EditText
 import androidx.core.widget.doOnTextChanged
 import br.com.zup.beagle.annotation.RegisterWidget
-import br.com.zup.beagle.interfaces.StateChangeable
-import br.com.zup.beagle.interfaces.WidgetState
 import br.com.zup.beagle.sample.utils.MaskApplier
-import br.com.zup.beagle.state.Observable
 import br.com.zup.beagle.widget.form.InputWidget
 
 enum class TextFieldInputType {
@@ -41,11 +38,7 @@ data class TextField(
     val color: String = "#000000",
     val mask: String? = null,
     val inputType: TextFieldInputType? = null
-) : InputWidget(), StateChangeable {
-
-    private val stateObservable = Observable<WidgetState>()
-
-    override fun getState(): Observable<WidgetState> = stateObservable
+) : InputWidget() {
 
     private lateinit var textFieldView: EditText
 
@@ -54,7 +47,7 @@ data class TextField(
         bind()
 
         this.doOnTextChanged { text, _, _, _ ->
-            stateObservable.notifyObservers(WidgetState(text.toString()))
+            notifyObservers(text.toString())
         }
     }
 
@@ -87,5 +80,5 @@ data class TextField(
         }
     }
 
-    fun getColorWithHashTag(value: String): String = if (value.startsWith("#")) value else "#$value"
+    private fun getColorWithHashTag(value: String): String = if (value.startsWith("#")) value else "#$value"
 }
